plugins: Handle more key events to improve usability in the Developer Console.
Before this commit, the user had to put the cursor manually at an
editable area (whichever part of the Gtk.TextView after the prompt)
to be able to type something. That could be confusing because it
gave the idea that the console got frozen.

This commit adds the following features:

- Input a text to the console even if the cursor is at the
  non-editable area.
- Delete the text when the backspace key is pressed while the
  cursor is at the non-editable area.
- Give freedom to up, left, right, down keys to move around
  the non-editable area if the cursor is in that area.